--- app/src/main/java/com/my/vitamateapp/RegistrationCompleteActivity.kt
@@ -1,14 +1,0 @@
-package com.my.vitamateapp
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class RegistrationCompleteActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_registration_complete)
-
-
-
-    }
-}
