+++ app/src/main/java/com/my/vitamateapp/registerPage/InBodyRecordActivity.kt
@@ -1,19 +1,21 @@
-package com.my.vitamateapp
+package com.my.vitamateapp.registerPage
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.my.vitamateapp.databinding.ActivityNoInBodyRecordBinding
+import com.my.vitamateapp.R
+import com.my.vitamateapp.databinding.ActivityInBodyRecordBinding
 
-class NoInBodyRecordActivity : AppCompatActivity() {
+class InBodyRecordActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityNoInBodyRecordBinding
+    private lateinit var binding: ActivityInBodyRecordBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_no_in_body_record)
+        setContentView(R.layout.activity_in_body_record)
 
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_no_in_body_record)
+
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_in_body_record)
 
         binding.nextBtn.setOnClickListener{
             startActivity(Intent(this, RegistrationCompleteActivity::class.java))
@@ -25,7 +27,6 @@ class NoInBodyRecordActivity : AppCompatActivity() {
             finish() // 현재 액티비티 종료..
         }
 
+
     }
-
-
 }
