+++ app/src/main/java/com/my/vitamateapp/registerPage/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.my.vitamateapp
+package com.my.vitamateapp.registerPage
 
 import android.content.ContentValues.TAG
 import android.content.Intent
@@ -10,6 +10,8 @@ import com.kakao.sdk.auth.model.OAuthToken
 import com.kakao.sdk.common.model.ClientError
 import com.kakao.sdk.common.model.ClientErrorCause
 import com.kakao.sdk.user.UserApiClient
+import com.my.vitamateapp.HomeActivity
+import com.my.vitamateapp.R
 import com.my.vitamateapp.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -21,15 +23,16 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
 
-        //토큰 보유 정보 확인.
+        // 로그인 전 토큰 정보 보기
         UserApiClient.instance.accessTokenInfo { tokenInfo, error ->
             if (error != null) {
-                Log.e(TAG, "토큰 정보 확인 실패", error)
-            } else if (tokenInfo != null) {
-                Log.i(TAG, "토큰 정보 확인 성공" +
+                Log.e(TAG, "토큰 정보 보기 실패", error)
+
+            }
+            else if (tokenInfo != null) {
+                Log.i(TAG, "토큰 정보 보기 성공" +
                         "\n회원번호: ${tokenInfo.id}" +
                         "\n만료시간: ${tokenInfo.expiresIn} 초")
-                // 토큰이 유효한 경우 로그인 없이 홈 화면으로 이동
                 gotoHome()
             }
         }
@@ -43,20 +46,8 @@ class MainActivity : AppCompatActivity() {
     //로그아웃 하고 다시 회원가입하지 않게 수정하기.
 
 
+
     private fun loginWithKakao(){
-
-        // 토큰 정보 보기, 토큰 정보 있으면 로그인 없이 HOME화면으로 바로가기
-        UserApiClient.instance.accessTokenInfo { tokenInfo, error ->
-            if (error != null) {
-                Log.e(TAG, "토큰 정보 보기 실패", error)
-            }
-            else if (tokenInfo != null) {
-                Log.i(TAG, "토큰 정보 보기 성공")
-                gotoHome()
-            }
-        }
-
-
 
         // 카카오계정으로 로그인 공통 callback 구성
         // 카카오톡으로 로그인 할 수 없어 카카오계정으로 로그인할 경우 사용됨
@@ -65,11 +56,10 @@ class MainActivity : AppCompatActivity() {
                 Log.e(TAG, "카카오계정으로 로그인 실패", error)
             } else if (token != null) {
                 Log.i(TAG, "카카오계정으로 로그인 성공 ${token.accessToken}")
-                gotoRegister()//Home화면으로 이동
+                gotoRegister()//회원정보입력화면으로 이동
 
             }
         }
-
 
         // 카카오톡이 설치되어 있으면 카카오톡으로 로그인, 아니면 카카오계정으로 로그인
         if (UserApiClient.instance.isKakaoTalkLoginAvailable(this)) {
