+++ app/src/main/java/com/my/vitamateapp/registerPage/NoInBodyRecordActivity.kt
@@ -1,20 +1,20 @@
-package com.my.vitamateapp
+package com.my.vitamateapp.registerPage
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.my.vitamateapp.databinding.ActivityInBodyRecordBinding
+import com.my.vitamateapp.R
+import com.my.vitamateapp.databinding.ActivityNoInBodyRecordBinding
 
-class InBodyRecordActivity : AppCompatActivity() {
+class NoInBodyRecordActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityInBodyRecordBinding
+    private lateinit var binding: ActivityNoInBodyRecordBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_in_body_record)
+        setContentView(R.layout.activity_no_in_body_record)
 
-
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_in_body_record)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_no_in_body_record)
 
         binding.nextBtn.setOnClickListener{
             startActivity(Intent(this, RegistrationCompleteActivity::class.java))
@@ -26,6 +26,7 @@ class InBodyRecordActivity : AppCompatActivity() {
             finish() // 현재 액티비티 종료..
         }
 
-
     }
+
+
 }
