+++ app/src/main/res/layout/activity_in_body_check.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".InBodyCheckActivity">
+    tools:context=".registerPage.InBodyCheckActivity">
 
     <TextView
       android:layout_width="wrap_content"
