+++ app/src/main/res/layout/activity_no_in_body_record.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".RegisterActivity">
+    tools:context=".registerPage.RegisterActivity">
 
 
     <TextView
