+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
   <string name="app_name">VitaMateApp</string>
+  <!-- TODO: Remove or change this placeholder text -->
+  <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
