+++ gradle.properties
@@ -22,8 +22,7 @@ kotlin.code.style=official
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
 
-# Gradle ?? ???
+
 org.gradle.caching=true
 
-# ?? ???? ??
 org.gradle.caching.directory=./.gradle/caches
